Relocate settings_tab and ui_theme into modules/
Merge commit bringing in PR #9 left the two new files at repo root;
this moves them into modules/ to match the reorg layout and updates
imports accordingly (main launcher imports from modules.*, and
settings_tab's internal import of ui_theme uses from modules import).
